Refactored common functionality to archive library.
- Moved common functionality for Test Harness and its dependencies into
common-lib to promote code reusability, ease of maintenance and cleanliness.
common-lib is now a custom version that contains the common-lib plus
HEBench common code.
