+++ common-lib/CMakeLists.txt
@@ -9,8 +9,8 @@
 
 cmake_minimum_required(VERSION 3.12)
 
-project(common-lib VERSION 0.1.4 LANGUAGES C CXX)
-set(ACTIVE_TARGET hebench_common-lib)
+project(hebench_common-lib VERSION 0.1.5 LANGUAGES C CXX)
+set(ACTIVE_TARGET ${PROJECT_NAME})
 
 set(HEADERS
     modules/general/include/pool.h
@@ -18,6 +18,8 @@ set(HEADERS
     modules/general/include/error.h
     modules/general/include/memory_buffer.h
     modules/general/include/api_decl.h
+    modules/general/include/hebench_utilities.h
+    modules/general/include/hebench_math_utils.h
     modules/args_parser/include/args_parser.h
     modules/config_reader/include/config_reader.h
     modules/logging/include/logging.h
@@ -34,6 +36,8 @@ set(SOURCES
     modules/args_parser/src/args_parser.cpp
     modules/config_reader/src/config_reader.cpp
     modules/threading/src/threading.cpp
+    modules/general/src/hebench_utilities.cpp
+    modules/general/src/hebench_math_utils.cpp
     modules/general/src/memory_buffer.cpp
     modules/logging/src/logging.cpp
     )
@@ -56,5 +60,5 @@ set_target_properties(${ACTIVE_TARGET} PROPERTIES
     CXX_EXTENSIONS NO
 )
 
-target_link_libraries(hebench_common-lib Threads::Threads)
-target_compile_options(hebench_common-lib PRIVATE -Wall -Wextra)
+target_link_libraries(${ACTIVE_TARGET} Threads::Threads)
+target_compile_options(${ACTIVE_TARGET} PRIVATE -Wall -Wextra -fPIC)
