+++ report_gen/report_gen_lib/src/hebench_report.cpp
@@ -10,7 +10,8 @@
 
 #include "hebench_report.h"
 #include "hebench_report_impl.h"
-#include "hebench_report_utils.h"
+#include "modules/general/include/hebench_math_utils.h"
+#include "modules/general/include/hebench_utilities.h"
 
 namespace hebench {
 namespace TestHarness {
