+++ test_harness/benchmarks/MatrixMultiply/latency/src/hebench_matmult_l.cpp
@@ -16,7 +16,7 @@
 
 #include "hebench/api_bridge/api.h"
 #include "include/hebench_engine.h"
-#include "include/hebench_math_utils.h"
+#include "modules/general/include/hebench_math_utils.h"
 
 #include "../include/hebench_matmult_l.h"
