+++ test_harness/benchmarks/categories/src/hebench_benchmark_latency.cpp
@@ -14,7 +14,7 @@
 
 #include "hebench/api_bridge/api.h"
 #include "include/hebench_engine.h"
-#include "include/hebench_math_utils.h"
+#include "modules/general/include/hebench_math_utils.h"
 
 #include "../include/hebench_benchmark_latency.h"
