+++ test_harness/src/hebench_config.cpp
@@ -19,8 +19,9 @@
 #include "include/hebench_config.h"
 #include "include/hebench_engine.h"
 #include "include/hebench_ibenchmark.h"
-#include "include/hebench_math_utils.h"
-#include "include/hebench_utilities.h"
+#include "include/hebench_utilities_harness.h"
+#include "modules/general/include/hebench_math_utils.h"
+#include "modules/general/include/hebench_utilities.h"
 
 namespace hebench {
 namespace Utilities {
